rename deploy to deployctl (#211)
* rename deploy to deployctl

renaming deploy to deployctl

Test Plan:
Pass: Run the make command
Pass: Docker-build
Pass: Make tools

deployctl is generated in bin/ path

JIRA: CGTS-47703

* rename deploy to deployctl

renaming deploy to deployctl

Test Plan:
Pass: Run the make command
Pass: Docker-build
Pass: Make tools

deployctl is generated in bin/ path


---------

(cherry picked from commit 443baa59cc4d4f1688152d7e214e1bfa4184b1f8)
