+++ cmd/deployctl/main.go
@@ -3,7 +3,7 @@
 
 package main
 
-import "github.com/wind-river/cloud-platform-deployment-manager/cmd/deploy/cmd"
+import "github.com/wind-river/cloud-platform-deployment-manager/cmd/deployctl/cmd"
 
 func main() {
 	cmd.Execute()
